// src/components/saved-property/auth/screen.enum.ts
export enum SavedPropertyAuthScreenEnum {
  /**
   * The short login screen
   */
  default = 'default',

  /**
   * Login screen
   */
  login = 'login',

  /**
   * Registration screen
   */
  registration = 'registration',

  /**
   * Forgot password screen
   */
  forgotPassword = 'forgot-password',
}